refactor: 앱 영화 타입 domain 타입 파일로 이동
