+++ src/types/domain.d.ts
@@ -1,3 +1,10 @@
+export interface AppMovie {
+  id: number;
+  title: string;
+  posterPath: string | null;
+  rating: number;
+}
+
 export interface RatedMovie {
   id: number;
   score: string;
